+++ scripts/moduleConfigs/cmsModules/nuxtStudio/fileConfig.js
@@ -1,5 +1,0 @@
-import { field, group } from '@nuxthq/studio/theme'
-
-export default defineNuxtSchema({
-    // ...
-})