+++ scripts/moduleConfigs/cmsModules/nuxtStudio/nuxtConfig.ts
@@ -1,5 +1,0 @@
-export default {
-    studio: {
-        enabled: true,
-    }
-}